+++ src/honeyhive/__init__.py
@@ -27,7 +27,7 @@ from .utils.config import config
 from .utils.dotdict import DotDict
 from .utils.logger import HoneyHiveLogger, get_logger
 
-__version__ = "0.1.0"
+__version__ = "0.1.0rc1"
 
 __all__ = [
     "HoneyHive",